requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.31.0

# Database
supabase>=2.3.4
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Data Processing
pandas>=2.2.0
numpy>=1.26.3
narwhals>=2.15.0
pyarrow>=15.0.0

# Visualization
plotly>=5.18.0
pydeck>=0.9.1
altair>=6.0.0

# AI/ML
requests>=2.31.0
transformers>=4.37.0
torch>=2.2.0

# Utilities
python-dotenv>=1.0.0
tenacity>=9.0.0
attrs>=23.2.0
urllib3>=2.0.0